Build the benchmarks on machines without a GPU (#187)
The benchmarks only built when CHUCKY_ENABLE_GPU was on, so a CPU-only
checkout had none. bench/ is now added unconditionally.

The GPU pipeline sits behind bench/bench_gpu.h, which has no CUDA in it.
bench_gpu.cuda.c forwards to tile_stream_gpu; bench_gpu.none.c stands in
when GPU support is off, and bench_util.c no longer reaches for cuda.h
or stream.gpu.h. The two-stream benchmark needs a GPU, so it is built
only where there is one, the way the GPU-only tests already are.
Without GPU support the --backend default is cpu, and asking for
--backend gpu fails with a message to reconfigure.

The batch-size target in the stream configuration is now 64 bits rather
than 32, so a batch above 4 GiB works and nothing has to be narrowed on
the way in. The benchmark command line reads sizes with fixed-width
types throughout, and --chunk-bytes, --batch-bytes, --memory-budget and
--backpressure now reject text they cannot read instead of parsing it as
something else. --batch-bytes also reaches runs that skip the auto-fit
path, which ignored it.

Closes #97.

---------

+++ bench/bench_util.h
@@ -8,7 +8,6 @@
 #include "sink_discard.h"
 #include "sink_metering.h"
 #include "sink_throttled.h"
-#include "stream.gpu.h"
 #include "test_data.h"
 #include "types.codec.h"
 #include "types.lod.h"
@@ -34,15 +33,15 @@ struct bench_config
   enum lod_reduce_method reduce_method;
   enum lod_reduce_method append_reduce_method;
   enum bench_backend backend;
-  enum dtype dtype;          // element type (default dtype_u16)
-  const int* chunk_ratios;   // power-of-2 distribution ratios; see
-                             // dims_budget_chunk_size for the -1/0/>0
-                             // conventions
-  size_t target_chunk_bytes; // 0 = use 1MB default
-  size_t min_chunk_bytes;    // auto-fit floor; 0 = no floor
-  size_t target_batch_bytes; // 0 = 512 MiB default; controls auto-K
-  size_t memory_budget;      // 0 = auto-detect
-  size_t min_shard_bytes;    // minimum uncompressed bytes per shard
+  enum dtype dtype;            // element type (default dtype_u16)
+  const int* chunk_ratios;     // power-of-2 distribution ratios; see
+                               // dims_budget_chunk_size for the -1/0/>0
+                               // conventions
+  size_t target_chunk_bytes;   // 0 = use 1MB default
+  size_t min_chunk_bytes;      // auto-fit floor; 0 = no floor
+  uint64_t target_batch_bytes; // 0 = 512 MiB default; controls auto-K
+  size_t memory_budget;        // 0 = auto-detect
+  size_t min_shard_bytes;      // minimum uncompressed bytes per shard
   uint32_t
     target_concurrent_shards; // cap on inner shard product (active files)
   uint32_t min_append_shards; // require at least N shards along the outer
@@ -77,12 +76,11 @@ struct bench_spec
   uint32_t min_append_shards; // 0 = no minimum (see bench_config)
 };
 
-// CLI driver: parses --fill, --codec, --reduce, --dtype, --frames, --json,
-// -o flags, inits CUDA, calls run_bench, handles xor_pattern_init/free.
+// Parse the shared bench flags and run one benchmark. The backend defaults to
+// gpu, or to cpu in a build without GPU support.
 int
 bench_stream_main(int ac, char* av[], struct bench_spec spec);
 
-// Two-stream variant: creates two GPU pipelines on the same CUDA context and
-// interleaves writer_append calls for balanced GPU utilisation.
+// Run two pipelines on one GPU, appending to each in turn.
 int
 bench_two_streams_main(int ac, char* av[], struct bench_spec spec);
